agent/requirements.txt
# LiveKit Agents Framework
livekit-agents>=0.8.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-google>=0.6.0
livekit-plugins-silero>=0.6.0

# Memory & Vector Store
mem0ai>=0.1.0
qdrant-client>=1.9.0

# Google Gemini
google-generativeai>=0.5.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.3.0
